+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/graph/GraphSource.java
@@ -33,6 +33,11 @@ import java.util.Set;
 public abstract class GraphSource implements Serializable {
 
     /**
+     * Start IP
+     */
+    protected int startIp = 0;
+    
+    /**
      * Gets the size of the graph source
      *
      * @return The size of the graph source
@@ -84,7 +89,7 @@ public abstract class GraphSource implements Serializable {
      * @param pos Position of the instruction
      * @return Instruction as string
      */
-    public abstract String insToString(int pos);
+    public abstract String insToString(int pos);        
 
     /**
      * Visits the code
@@ -162,7 +167,7 @@ public abstract class GraphSource implements Serializable {
         for (int i = 0; i < siz; i++) {
             refs.put(i, new ArrayList<>());
         }
-        visitCode(0, 0, refs, -1);
+        visitCode(startIp, 0, refs, -1);
         int pos = 0;
         for (int e : alternateEntries) {
             pos++;